Ajax Included

+++ TodoWeb/views.py
@@ -1,6 +1,7 @@
 from django.shortcuts import render,redirect
 from django.http import HttpResponse
 from .models import Task
+import json
 
 def Home(request):
     Data = Task.objects.all()
@@ -9,46 +10,91 @@ def Home(request):
             }
     return render(request,'JavaScript.html', DataDictionary)
 
-def NewTask(request):
+def NewTask(request, UrlValue):
     if request.method == 'POST':
-        TaskContent = request.POST.get('TaskContent')
-        if(TaskContent != ""):
-            NewTaskObeject = Task(Content = TaskContent)
-            NewTaskObeject.save()
-            Data = Task.objects.all()
-            DataDictionary = {
-                        'Data' :Data 
-                    } 
-            return redirect('Home')
-        else:
-            TaskContent = request.POST.get('TaskContent')
-            Data = Task.objects.all()
-            DataDictionary = {
-                        'Data' : Data 
-                    } 
-            return redirect('Home')
-def StatusChange(request, id):
-    SpecificTask = Task.objects.get(pk = id)
-    if SpecificTask.Status == 'true':
-        SpecificTask.Status = 'false'
-        SpecificTask.save()
-        return redirect('Home')
-    else:
-        SpecificTask.Status = 'true'
-        SpecificTask.save()
-        return redirect('Home')  
+        if(UrlValue != ""):
+            
+            IdentifyUrl = UrlValue
 
-def TaskEdit(request, id, ChangedContent):
-    SpecificTask = Task.objects.get(pk = id)
-    SpecificTask.Content = ChangedContent
-    SpecificTask.save()
-    return redirect('Home')
-       
-def TaskDelete(request, id):
-    SpecificTask = Task.objects.get(pk = id)
-    SpecificTask.delete()
-    return redirect('Home')
-        
+            if (IdentifyUrl[0] == 'N'):
+                    
+                RealData = IdentifyUrl[2:]   
 
+                NewTaskObeject = Task(Content = RealData)
 
+                NewTaskObeject.save()
+  
+                ResponsedData = {"TaskId" :NewTaskObeject.id, "TaskData":NewTaskObeject.Content}
 
+                Response = json.dumps(ResponsedData)
+     
+                return HttpResponse(Response)
+
+            elif (IdentifyUrl[0] == 'C'):
+
+                RealData = IdentifyUrl[2:]  
+
+                TaskCompleteIdentity = int(RealData) 
+
+                SpecificTask = Task.objects.get(pk = TaskCompleteIdentity)
+
+                if SpecificTask.Status == 'true':
+
+                    SpecificTask.Status = 'false'
+
+                    SpecificTask.save()
+
+                    TaskCompleteIdentity = str(TaskCompleteIdentity)
+
+                    ResponsedData = {"TaskId" :TaskCompleteIdentity, "TaskStatus":'False'}
+
+                    Response = json.dumps(ResponsedData)
+     
+                    return HttpResponse(Response)
+
+                else:
+
+                    SpecificTask.Status = 'true'
+
+                    SpecificTask.save()
+
+                    TaskCompleteIdentity = str(TaskCompleteIdentity)
+
+                    ResponsedData = {"TaskId" :TaskCompleteIdentity, "TaskStatus":'True'}
+
+                    Response = json.dumps(ResponsedData)
+     
+                    return HttpResponse(Response) 
+
+            elif (IdentifyUrl[0] == 'U'):
+
+                RealData = IdentifyUrl[2:]
+
+                DataList = RealData.split("-") 
+
+                SpecificTask = Task.objects.get(pk = int(DataList[0]))
+
+                SpecificTask.Content = DataList[1]
+
+                SpecificTask.save()
+
+                ResponsedData = {"TaskId" :DataList[0], "TaskNewContent":DataList[1]}
+
+                Response = json.dumps(ResponsedData)
+     
+                return HttpResponse(Response)
+
+            else:
+
+                RealData = IdentifyUrl[2:]
+
+                SpecificTask = Task.objects.get(pk = int(RealData))
+
+                SpecificTask.delete()
+
+                ResponsedData = {"TaskId" : RealData}
+
+                Response = json.dumps(ResponsedData)
+     
+                return HttpResponse(Response)
+
